Split comments to own controller
