+++ src/test/java/org/ambraproject/wombat/controller/CommentControllerTest.java
@@ -22,17 +22,13 @@
 
 package org.ambraproject.wombat.controller;
 
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import org.ambraproject.wombat.config.SpringMvcConfiguration;
 import org.ambraproject.wombat.identity.ArticlePointer;
 import org.ambraproject.wombat.identity.RequestedDoiVersion;
-import org.ambraproject.wombat.model.Reference;
 import org.ambraproject.wombat.service.*;
 import org.ambraproject.wombat.service.remote.ArticleApi;
 import org.ambraproject.wombat.service.remote.CorpusContentApi;
-import org.ambraproject.wombat.service.remote.orcid.OrcidApi;
-import org.ambraproject.wombat.service.remote.orcid.OrcidApiImpl;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.test.annotation.DirtiesContext;
@@ -51,23 +47,18 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
 
 @ContextConfiguration(
-    classes = {SpringMvcConfiguration.class, ArticleController.class, ArticleControllerTest.class})
+    classes = {SpringMvcConfiguration.class, CommentController.class, CommentControllerTest.class})
 @Configuration
 @DirtiesContext(classMode = ClassMode.AFTER_EACH_TEST_METHOD)
-public class ArticleControllerTest extends ControllerTest {
+public class CommentControllerTest extends ControllerTest {
 
   private static final String EXPECTED_DOI = "10.1371/journal.pbio.1001091";
 
   private static final int EXPECTED_INGESTION_NUMBER = 2;
-
-  @Bean
-  protected ArticleResolutionService articleResolutionService() {
-    final ArticleResolutionService articleResolutionService = mock(ArticleResolutionService.class);
-    return articleResolutionService;
-  }
 
   @Bean
   protected ArticleService articleService() {
@@ -99,44 +90,32 @@ public class ArticleControllerTest extends ControllerTest {
   }
 
   @Bean
-  protected CitationDownloadService citationDownloadService() {
-    final CitationDownloadService citationDownloadService = mock(CitationDownloadService.class);
-    return citationDownloadService;
+  protected ArticleResolutionService articleResolutionService() {
+    final ArticleResolutionService articleResolutionService = mock(ArticleResolutionService.class);
+    return articleResolutionService;
   }
 
   @Bean
-  protected ParseReferenceService parseReferenceService() {
-    final ParseReferenceService parseReferenceService = new ParseReferenceService();
-    return parseReferenceService;
+  protected CommentValidationService commentValidationService() {
+    final CommentValidationService commentValidationService = mock(CommentValidationService.class);
+    return commentValidationService;
   }
 
   @Bean
-  protected ParseXmlService parseXmlService() {
-    final ParseXmlService parseXmlService = spy(ParseXmlServiceImpl.class);
-    return parseXmlService;
+  protected CommentService commentService() {
+    final CommentService commentService = mock(CommentService.class);
+    return commentService;
   }
 
-  @Bean
-  protected DoiToJournalResolutionService doiToJournalResolutionService() {
-    final DoiToJournalResolutionService doiToJournalResolutionService =
-        spy(DoiToJournalResolutionService.class);
-    return doiToJournalResolutionService;
-  }
-
-  @Bean
-  protected OrcidApi orcidApi() {
-    final OrcidApi orcidApi = spy(OrcidApiImpl.class);
-    return orcidApi;
-  }
 
   /**
-   * Test successful rendering of an article.
+   * Test successful rendering of an article's comments.
    *
    * @throws URISyntaxException if an invalid URI
-   * @throws Exception if failed to render article
+   * @throws Exception if failed to render
    */
   @Test
-  public void testRenderArticleShouldSucceed() throws URISyntaxException, Exception {
+  public void testRenderArticleCommentsShouldSucceed() throws URISyntaxException, Exception {
     final RequestedDoiVersion expectedRequestedDoi = RequestedDoiVersion.of(EXPECTED_DOI);
 
     final ArticlePointer expectedArticlePointer =
@@ -144,7 +123,7 @@ public class ArticleControllerTest extends ControllerTest {
             expectedRequestedDoi, EXPECTED_DOI, EXPECTED_INGESTION_NUMBER,
             expectedRequestedDoi.getRevisionNumber());
 
-    final ArticleResolutionService mockArticleResolutionService = 
+    final ArticleResolutionService mockArticleResolutionService =
         applicationContext.getBean(ArticleResolutionService.class);
     when(mockArticleResolutionService.toIngestion(expectedRequestedDoi))
         .thenReturn(expectedArticlePointer);
@@ -163,34 +142,57 @@ public class ArticleControllerTest extends ControllerTest {
     doReturn(ingestionMetadata, relationships).when(mockArticleApi)
         .requestObject(any(), eq(Map.class));
 
-    final ArticleMetadata mockArticleMetadata = applicationContext.getBean(ArticleMetadata.class);
-    when(mockArticleMetadata.getArticlePointer()).thenReturn(expectedArticlePointer);
-
-    final String expectedHtml =
-        "<html><title>This is a title</title><body>This is the body</body></html>";
-    final Reference reference =
-        Reference.build().setTitle("Reference title").setPublisherName("Publisher Name")
-            .setAuthors(ImmutableList.of()).setCollabAuthors(ImmutableList.of()).build();
-    final ImmutableList<Reference> expectedReferences = ImmutableList.of(reference);
-    final ArticleController.XmlContent expectedXmlContent =
-        new ArticleController.XmlContent(expectedHtml, expectedReferences);
-
-    final CorpusContentApi mockCorpusContentApi =
-        applicationContext.getBean(CorpusContentApi.class);
-    when(mockCorpusContentApi.readManuscript(any(), any(), any(), any()))
-        .thenReturn(expectedXmlContent);
-
-    final String expectedViewName = NOSPACE_JOINER.join(DESKTOP_PLOS_ONE, "/ftl/article/article");
-    final String requestUri = NOSPACE_JOINER.join("/article?id=", EXPECTED_DOI);
+    final String expectedViewName = NOSPACE_JOINER.join(DESKTOP_PLOS_ONE, "/ftl/article/comment/comments");
+    final String requestUri = NOSPACE_JOINER.join("/article/comments?id=", EXPECTED_DOI);
     mockMvc.perform(get(new URI(requestUri))).andExpect(status().isOk())
-        .andExpect(model().attribute("articleText", expectedHtml))
-        .andExpect(model().attribute("references", expectedReferences))
         .andExpect(view().name(expectedViewName));
 
     verify(mockArticleResolutionService).toIngestion(expectedRequestedDoi);
     verify(mockArticleService).getItemTable(expectedArticlePointer);
     verify(mockArticleApi, times(2)).requestObject(any(), eq(Map.class));
-    verify(mockArticleMetadata).getArticlePointer();
-    verify(mockCorpusContentApi).readManuscript(any(), any(), any(), any());
+  }
+
+  /**
+   * Test successful rendering of an article's comment form.
+   *
+   * @throws URISyntaxException if an invalid URI
+   * @throws Exception if failed to render
+   */
+  @Test
+  public void testRenderArticleCommentFormShouldSucceed() throws URISyntaxException, Exception {
+    final RequestedDoiVersion expectedRequestedDoi = RequestedDoiVersion.of(EXPECTED_DOI);
+
+    final ArticlePointer expectedArticlePointer =
+        new ArticlePointer(
+            expectedRequestedDoi, EXPECTED_DOI, EXPECTED_INGESTION_NUMBER,
+            expectedRequestedDoi.getRevisionNumber());
+
+    final ArticleResolutionService mockArticleResolutionService =
+        applicationContext.getBean(ArticleResolutionService.class);
+    when(mockArticleResolutionService.toIngestion(expectedRequestedDoi))
+        .thenReturn(expectedArticlePointer);
+
+    final Map<String, Object> itemResponse = ImmutableMap.of("items", ImmutableMap.of());
+    final ArticleService mockArticleService = applicationContext.getBean(ArticleService.class);
+    doAnswer(invocation -> {
+      return itemResponse;
+    }).when(mockArticleService).getItemTable(expectedArticlePointer);
+
+    final ImmutableMap<String, String> journal = ImmutableMap.of("journalKey", DESKTOP_PLOS_ONE);
+    final ImmutableMap<String, Object> ingestionMetadata = ImmutableMap.of("journal", journal);
+    final ImmutableMap<String, List<Map<String, ?>>> relationships = ImmutableMap.of();
+
+    final ArticleApi mockArticleApi = applicationContext.getBean(ArticleApi.class);
+    doReturn(ingestionMetadata, relationships).when(mockArticleApi)
+        .requestObject(any(), eq(Map.class));
+
+    final String expectedViewName = NOSPACE_JOINER.join(DESKTOP_PLOS_ONE, "/ftl/article/comment/newComment");
+    final String requestUri = NOSPACE_JOINER.join("/article/comments/new?id=", EXPECTED_DOI);
+    mockMvc.perform(get(new URI(requestUri))).andExpect(status().isOk())
+        .andExpect(view().name(expectedViewName));
+
+    verify(mockArticleResolutionService).toIngestion(expectedRequestedDoi);
+    verify(mockArticleService).getItemTable(expectedArticlePointer);
+    verify(mockArticleApi, times(2)).requestObject(any(), eq(Map.class));
   }
 }
